requirements.txt
flask==2.0.1
Pillow==9.5.0
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose==3.3.0